add splash screen for unathorized user
